+++ testapps/GHIssues/src/main/java/ghissues/gh1069/Gh1069User.java
@@ -1,20 +1,20 @@
-package test;
+package ghissues.gh1069;
 
 import act.db.jpa.JPADao;
 
 import javax.persistence.*;
 
 @Entity
-public class User {
+public class Gh1069User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     public Integer id;
 
-    @Unique(entity = User.class, field = "name")
+    @Unique(entity = Gh1069User.class, field = "name")
     public String name;
 
-    public static class Dao extends JPADao<Integer, User> {
+    public static class Dao extends JPADao<Integer, Gh1069User> {
     }