+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ActFramework Change Log
 
 **1.8.20**
+* Error encountered with customised Unique validator when working with JPA #1069
 * Act-Test - add refId property to Scenario #1091
 * Add switch to disable auto hot deploy in DEV mode #1090
 * GH 1078 - Log warn message when `@Global` is found on instance method of an abstract class